helpers/opsi/src/requirements.txt
requests
pandas
grpcio
grpcio-tools